+++ lib/bullet_train.dart
@@ -3,3 +3,4 @@ library bullet_train;
 export 'src/bullet_train_client.dart';
 export 'src/bullet_train_config.dart';
 export 'src/model/index.dart';
+export 'src/store/index.dart';