+++ lib/src/bullet_train_config.dart
@@ -11,6 +11,8 @@ class BulletTrainConfig {
   final int receiveTimeout;
   final int sendTimeout;
 
+  final bool usePersistantStorage;
+
   /// Bullet train config initialization
   /// change only if you have self-hosted bullet train
   /// [baseURI], [flagsURI], [identitiesURI], [traitsURI]
@@ -18,15 +20,15 @@ class BulletTrainConfig {
   /// Connection settings timeouts in milliseconds
   /// [connectionTimeout], [connectTimeout], [receiveTimeout], [sendTimeout]
 
-  const BulletTrainConfig({
-    this.baseURI = 'https://api.bullet-train.io/api/v1/',
-    this.flagsURI = 'flags/',
-    this.identitiesURI = 'identities/',
-    this.traitsURI = 'traits/',
-    this.connectTimeout = 2000,
-    this.receiveTimeout = 5000,
-    this.sendTimeout = 5000,
-  });
+  const BulletTrainConfig(
+      {this.baseURI = 'https://api.bullet-train.io/api/v1/',
+      this.flagsURI = 'flags/',
+      this.identitiesURI = 'identities/',
+      this.traitsURI = 'traits/',
+      this.connectTimeout = 2000,
+      this.receiveTimeout = 5000,
+      this.sendTimeout = 5000,
+      this.usePersistantStorage = false});
 
   /// Client options from config
   BaseOptions get clientOptions => BaseOptions(