+++ lib/src/model/flag_and_traits.dart
@@ -9,6 +9,7 @@ part 'flag_and_traits.g.dart';
 /// Holds a list of feature flags and user traits.
 @freezed
 abstract class FlagAndTraits with _$FlagAndTraits {
+  @JsonSerializable(explicitToJson: true)
   const factory FlagAndTraits(
       {@Default(<Flag>[]) List<Flag> flags,
       @Default(<Trait>[]) List<Trait> traits}) = _FlagAndTraits;