+++ lib/src/model/index.dart
@@ -2,5 +2,6 @@ export 'feature.dart';
 export 'feature_user.dart';
 export 'flag.dart';
 export 'flag_and_traits.dart';
+export 'flag_type.dart';
 export 'model_utils.dart';
 export 'trait.dart';