+++ lib/src/store/crud_store.dart
@@ -2,10 +2,11 @@ import '../model/flag.dart';
 
 /// Abstract for CRUD operations over storage
 abstract class CrudStore<T extends Flag> {
-  Flag create(Flag flag);
-  Flag read(String id);
-  Flag update(Flag flag);
-  void delete(String id);
-  List<Flag> getAll();
-  void clear();
+  Future<void> create(T item);
+  Future<T> read(String id);
+  Future<void> update(T item);
+  Future<void> delete(T item);
+  Future<List<T>> getAll();
+  Future<void> clear();
+  Future<void> init();
 }